Analysis of variance

This is wrapper to the ez::ezANOVA() function.

ez_anova(
  data,
  dv,
  wid,
  within = NULL,
  within_full = NULL,
  within_covariates = NULL,
  between = NULL,
  between_covariates = NULL,
  observed = NULL,
  diff = NULL,
  reverse_diff = FALSE,
  type = 2,
  white.adjust = FALSE,
  detailed = FALSE,
  return_aov = FALSE
)

Arguments

data

Data frame containing the data to be analyzed.

dv

Name of the column in data that contains the dependent variable. Values in this column must be numeric.

wid

Name of the column in data that contains the variable specifying the case/Ss identifier. This should be a unique value per case/Ss.

within

Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss.

within_full

Same as within, but intended to specify the full within-Ss design in cases where the data have not already been collapsed to means per condition specified by within and when within only specifies a subset of the full design.

within_covariates

Names of columns in data that contain predictor variables that are manipulated (or observed) within-Ss and are to serve as covariates in the analysis.

between

Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss.

between_covariates

Names of columns in data that contain predictor variables that are manipulated (or observed) between-Ss and are to serve as covariates in the analysis.

observed

Names of columns in data that are already specified in either within or between that contain predictor variables that are observed variables (i.e. not manipulated).

diff

Names of any variables to collapse to a difference score. If a single value, may be specified by name alone; if multiple values, must be specified as a .() list.

reverse_diff

Logical. If TRUE, triggers reversal of the difference collapse requested by diff. Take care with variables with more than 2 levels.

type

Numeric value (either 1, 2 or 3) specifying the Sums of Squares type to employ when data are unbalanced (eg. when group sizes differ).

white.adjust

Only affects behaviour if the design contains only between-Ss predictor variables. If not FALSE, the value is passed as the white.adjust argument to Anova, which provides heteroscedasticity correction.

detailed

Logical. If TRUE, returns extra information (sums of squares columns, intercept row, etc.) in the ANOVA table.

return_aov

Logical. If TRUE, computes and returns an aov object corresponding to the requested ANOVA (useful for computing post-hoc contrasts).

Value

A list containing one or more components as returned by ez::ezANOVA().

Examples

ez_anova(data = selfesteem2_long,
            dv = score,
            wid = id,
            within = c(time, treatment),
            detailed = TRUE,
            return_aov = TRUE)
#> Warning: Converting "time" to factor for ANOVA.
#> $ANOVA
#>           Effect DFn DFd         SSn        SSd          F            p p<.05
#> 1    (Intercept)   1  11 527535.6806 4641.15278 1250.31275 1.112064e-12     *
#> 2           time   2  22    258.6944  103.97222   27.36922 1.075243e-06     *
#> 3      treatment   1  11    316.6806  224.15278   15.54068 2.303450e-03     *
#> 4 time:treatment   2  22    266.3611   96.30556   30.42371 4.629997e-07     *
#>          ges
#> 1 0.99048898
#> 2 0.04858771
#> 3 0.05883780
#> 4 0.04995572
#> 
#> $`Mauchly's Test for Sphericity`
#>           Effect         W          p p<.05
#> 2           time 0.4690776 0.02271033     *
#> 4 time:treatment 0.6160746 0.08874965      
#> 
#> $`Sphericity Corrections`
#>           Effect       GGe        p[GG] p[GG]<.05       HFe        p[HF]
#> 2           time 0.6532010 5.033982e-05         * 0.7054565 2.810626e-05
#> 4 time:treatment 0.7225823 1.254728e-05         * 0.8028383 4.815620e-06
#>   p[HF]<.05
#> 2         *
#> 4         *
#> 
#> $aov
#> 
#> Call:
#> aov(formula = formula(aov_formula), data = data)
#> 
#> Grand Mean: 85.59722
#> 
#> Stratum 1: id
#> 
#> Terms:
#>                 Residuals
#> Sum of Squares   4641.153
#> Deg. of Freedom        11
#> 
#> Residual standard error: 20.54076
#> 
#> Stratum 2: id:time
#> 
#> Terms:
#>                     time Residuals
#> Sum of Squares  258.6944  103.9722
#> Deg. of Freedom        2        22
#> 
#> Residual standard error: 2.173939
#> 2 out of 4 effects not estimable
#> Estimated effects may be unbalanced
#> 
#> Stratum 3: id:treatment
#> 
#> Terms:
#>                 treatment Residuals
#> Sum of Squares   316.6806  224.1528
#> Deg. of Freedom         1        11
#> 
#> Residual standard error: 4.514147
#> 2 out of 3 effects not estimable
#> Estimated effects are balanced
#> 
#> Stratum 4: id:time:treatment
#> 
#> Terms:
#>                 time:treatment Residuals
#> Sum of Squares       266.36111  96.30556
#> Deg. of Freedom              2        22
#> 
#> Residual standard error: 2.092254
#> Estimated effects may be unbalanced
#>