Self Esteem Score Data for Two-way Repeated Measures ANOVA

Data are the self esteem score of 12 individuals enrolled in 2 successive short-term trials (4 weeks) - control (placebo) and special diet trials.

The self esteem score was recorded at three time points: at the beginning (t1), midway (t2) and at the end (t3) of the trials.

The same 12 participants are enrolled in the two different trials with enough time between trials.

Two-way repeated measures ANOVA can be performed in order to determine whether there is interaction between time and treatment on the self esteem score.

This data set was taken from the datarium R package.

data("selfesteem2")

Format

A data frame with 24 rows and 5 columns.

Examples

data(selfesteem2)
selfesteem2
#> # A tibble: 24 × 5
#>    id    treatment    t1    t2    t3
#>    <fct> <fct>     <dbl> <dbl> <dbl>
#>  1 1     ctr          83    77    69
#>  2 2     ctr          97    95    88
#>  3 3     ctr          93    92    89
#>  4 4     ctr          92    92    89
#>  5 5     ctr          77    73    68
#>  6 6     ctr          72    65    63
#>  7 7     ctr          92    89    79
#>  8 8     ctr          92    87    81
#>  9 9     ctr          95    91    84
#> 10 10    ctr          92    84    81
#> # ℹ 14 more rows