Visual versus Verbal Perception and Responses

An experiment studying the interaction between visual versus perception and visual versus verbal responses.

vizverb

Format

A data frame with 80 observations on the following 5 variables.

subject

Subject identifying number (s1 to s20)

task

Describe a diagram (visual) or a sentence (verbal)

response

Point response (visual) or say response (verbal)

time

Response time (in seconds)

Source

This data set was taken from the Stats2Data R package. From the description in that package, the original data appear to have been collected in a Mount Holyoke College psychology class based replication of an experiment by Brooks, L., R. (1968) "Spatial and verbal components of the act of recall," Canadian J. Psych. V 22, pp. 349 - 368.

Details

Subjects carried out two kinds of tasks. One task was visual (describing a diagram), and the other was classed as verbal (reading and describing a sentence sentences). They reported the results either by pointing (a "visual" response), or speaking (a verbal response). Time to complete each task was recorded in seconds.